requirements.txt
aioredis==1.3.1
alabaster==0.7.12
amqp==2.5.2
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.2.3
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
auth0-python==3.10.0
Automat==20.2.0
autopep8==1.5
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
bleach==3.1.5
bokeh==2.0.2
boto3==1.14.2
botocore==1.17.2
bs4==0.0.1
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
cryptography==2.9.2
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.2
django-admin==2.0.0
django-allauth==0.42.0
django-authtools==1.7.0
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-excel-response2==2.0.8
django-filter==2.3.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-mpesa==0.1
django-mpesapy==0.1.2
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-storages==1.9.1
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
drf-yasg==1.17.1
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
Faker==0.9.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
h5py==2.10.0
hiredis==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.9
imagesize==1.2.0
incremental==17.5.0
inflection==0.5.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
kombu==4.6.8
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mixer==6.1.3
mlxtend==0.17.2
more-itertools==8.1.0
msgpack==0.6.2
mysql-connector-python==8.0.20
nbconvert==5.6.1
nbformat==5.0.6
nltk==3.5
notebook==6.0.3
numpy==1.18.3
oauthlib==3.1.0
openpyxl==3.0.3
packaging==20.1
pandas==1.0.3
pandocfilters==1.4.2
parse==1.15.0
parsel==1.6.0
parso==0.7.0
pdfminer==20191125
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
Protego==0.1.16
protobuf==3.11.3
psycopg2==2.8.5
psycopg2-binary==2.8.5
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pycryptodome==3.9.7
PyDispatcher==2.0.5
pyee==7.0.2
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyppeteer==0.2.2
pyquery==1.4.1
pyrsistent==0.16.0
pytesseract==0.3.4
pytest==5.3.4
pytest-cov==2.8.1
pytest-django==3.8.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-social-auth==0.3.6
python3-openid==3.1.0
qtconsole==4.7.3
QtPy==1.9.0
queuelib==1.5.0
regex==2020.5.14
requests==2.23.0
requests-html==0.10.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
scikit-learn==0.23.0
scipy==1.4.1
seaborn==0.10.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.14.0
sklearn==0.0
snowballstemmer==2.0.0
social-auth-core==3.2.0
socks==0
soupsieve==2.0.1
source==1.2.0
Sphinx==3.0.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.16
sqlparse==0.3.0
stripe==2.48.0
svgwrite==1.3.1
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.2
threadpoolctl==2.0.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
Tree==0.2.4
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.8
validate-email==1.3
vine==1.3.0
virtualenv==20.0.20
w3lib==1.22.0
wcwidth==0.1.8
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
whitenoise==5.1.0
widgetsnbextension==3.5.1
xlwt==1.3.0
zope.interface==5.1.0
